// tcl-web-app/src/app/models/interfaces/user-response.ts
// Generated by https://quicktype.io

export interface UserResponse {
  content:          User[];
  pageable:         Pageable;
  totalPages:       number;
  totalElements:    number;
  last:             boolean;
  size:             number;
  number:           number;
  sort:             Sort;
  first:            boolean;
  numberOfElements: number;
  empty:            boolean;
}

export interface User {
  uuid:              string;
  nick:              string;
  nombre:            string;
  fechaDeNacimiento: string;
  email:             string;
  role:              string;
  avatar:            string;
}

export interface Pageable {
  sort:       Sort;
  offset:     number;
  pageSize:   number;
  pageNumber: number;
  paged:      boolean;
  unpaged:    boolean;
}

export interface Sort {
  empty:    boolean;
  unsorted: boolean;
  sorted:   boolean;
}